% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_in.R
\name{centrality_in}
\alias{centrality_in}
\title{Title}
\usage{
centrality_in(cen_edges, n_nodes, all_edges_width)
}
\arguments{
\item{cen_edges}{edges df}

\item{n_nodes}{number of nodes within the network}

\item{all_edges_width}{sum of the width of all edges within the network}
}
\value{
DF with in_degree centrality measures
}
\description{
Title
}
