% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_out.R
\name{centrality_out}
\alias{centrality_out}
\title{Title centrality_out}
\usage{
centrality_out(cen_edges, n_nodes, all_edges_width)
}
\arguments{
\item{cen_edges}{edges}

\item{n_nodes}{number of nodes within the network}

\item{all_edges_width}{sum of the width of all edges within the network}
}
\value{
df with out_degree centrality for each node
}
\description{
Title centrality_out
}
