% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_sev.R
\name{centrality_sev}
\alias{centrality_sev}
\title{Title centrality_sev}
\usage{
centrality_sev(cen_edges, in_df, nodes_id_value, out_in_value)
}
\value{
A DF with severity weighted out_degree centrality measures: See Klintwall 2021
}
\description{
Title centrality_sev
}
