% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanAggregate.R
\name{pecanAggregate}
\alias{pecanAggregate}
\title{Title}
\usage{
pecanAggregate(
  data,
  nodes = NULL,
  edges = NULL,
  nodes_attribute = NULL,
  drop_nodes = NULL,
  drop_nodes_att = NULL,
  attribute_pattern = NULL,
  edges_sep = NULL
)
}
\arguments{
\item{edges_sep}{}
}
\description{
Title
}
