% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanCen.R
\name{pecanCen}
\alias{pecanCen}
\title{Title pecanCen}
\usage{
pecanCen(
  edges,
  nodes,
  centrality_by,
  sev_weighted = NULL,
  nodes_label = "auto"
)
}
\arguments{
\item{edges}{DF with edge columns "from", "to", "strength"}

\item{nodes}{DF with nodes columns "id", "value"}

\item{sev_weighted}{logical, should a severity weigthed out_degreee be calculated?}
}
\description{
Title pecanCen
}
