% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanExtractAgg.R
\name{pecanExtractAgg}
\alias{pecanExtractAgg}
\title{Title}
\usage{
pecanExtractAgg(
  data,
  nodes_include,
  edges_include,
  attributes_include = NULL,
  nodes_drop_NA0 = NULL,
  edges_drop_NA0 = NULL,
  attribute_drop_NA0 = NULL,
  edges_direction = "from_to"
)
}
\arguments{
\item{data}{}
}
\description{
Title
}
