% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanQGCP.R
\name{pecanQGCP}
\alias{pecanQGCP}
\title{Title pecanQGCP}
\usage{
pecanQGCP(
  edges,
  nodes,
  theme_bw = TRUE,
  print = TRUE,
  verbose = TRUE,
  weighted = TRUE,
  signed = TRUE,
  orderBy = "default",
  decreasing = FALSE,
  scale = "raw0",
  include = c("OutStrength", "InStrength", "Closeness", "Betweenness"),
  labels = "none"
)
}
\arguments{
\item{edges}{Edges df}

\item{nodes}{nodes df}

\item{theme_bw}{see centralityPlot function from qgraph}

\item{print}{see centralityPlot function from qgraph}

\item{verbose}{see centralityPlot function from qgraph}

\item{weighted}{see centralityPlot function from qgraph}

\item{signed}{see centralityPlot function from qgraph}

\item{orderBy}{see centralityPlot function from qgraph}

\item{decreasing}{see centralityPlot function from qgraph}

\item{scale}{see centralityPlot function from qgraph}

\item{include}{see centralityPlot function from qgraph}

\item{labelssee}{centralityPlot function from qgraph}
}
\value{
A centrality plot made with qgraph
}
\description{
Title pecanQGCP
}
