% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanQGCT.R
\name{pecanQGCT}
\alias{pecanQGCT}
\title{Title pecanQGCT}
\usage{
pecanQGCT(
  edges,
  nodes,
  standardized = TRUE,
  relative = FALSE,
  weighted = TRUE,
  signed = TRUE,
  labels = "none"
)
}
\arguments{
\item{edges}{Edges df}

\item{nodes}{Nodes df}

\item{standardized}{see centralityTable function from qgraph}

\item{relative}{see centralityTable function from qgraph}

\item{weighted}{see centralityTable function from qgraph}

\item{signed}{see centralityTable function from qgraph}

\item{labels}{see centralityTable function from qgraph}
}
\value{
A centrality table made with qgraph
}
\description{
Title pecanQGCT
}
