% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanRel.R
\name{pecanRel}
\alias{pecanRel}
\title{Title pecanRel}
\usage{
pecanRel(data_t1, data_t2, n_nodes)
}
\arguments{
\item{data_t1}{edges t1. "from", "to" and "width" are required}

\item{data_t2}{edges t2. "from", "to" and "width" are required}

\item{n_nodes}{number of nodes}
}
\value{
A df with different edge reliability measures Pearson_standard and Spearman_standard only
include edges which are eithe prensent in t1 or t2. Pearson_all_pe and Spearan_all_pe include all possible edges
}
\description{
Title pecanRel
}
