% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanVisEx.R
\name{pecanVisEx}
\alias{pecanVisEx}
\title{pecanVisEx}
\usage{
pecanVisEx(
  data,
  p_id,
  edges,
  nodes,
  edges_sep,
  label_in_names = FALSE,
  labels_sep = "none",
  labels = "none",
  regulation_by = "none",
  cut_by = "none",
  cut_off = "none",
  nodes_color = "red",
  edges_color = "black",
  unweigthed,
  max10 = 10,
  arrows = "to",
  shadow = "none",
  shadow_color = "default10",
  shadow_by = "none",
  use_ranking = "none",
  min_rank = "none",
  use_seed = "random"
)
}
\arguments{
\item{data}{df which includes the nodes and edges. Requires a column with the name "p_id"}

\item{p_id}{id of the person of interest...}

\item{edges}{Specify the columns of the edges e.g. 2:12. Edge names need to have a "to" and a "from" specification within their name and a separator.
Names can be numbers E.g 1_2 equals: "from" = 1 and "to" = 2 or words: Pain_Insomnia equals "from" = Pain and "to" = Insomnia}

\item{nodes}{Specify the columns of the nodes e.g. 13:17. Node names must match the edge names!}

\item{edges_sep}{How are "from" and "to" separated in the names?}

\item{label_in_names}{Logical. Are there additional labels in the node names. Must be separated and be second part of the name. e.g 1_Pain or Insomnia_I couldnt sleep}

\item{labels_sep}{How are the labels separated in the node names}

\item{labels}{Its possible to just add vector with the labels}

\item{regulation_by}{Should there be a regularization?
                     Interacts with the cut_off parameter.

                     Can be "nodes","number" or "value". Default is "none"

                     If nodes: Number of edges equals the number of nodes*cut_off

                     If number: The cut_off parameter sets the  number of edges in the network

                     If value: Edges under a specific cut_off are excluded from the network}

\item{cut_by}{Only variable that still need to be specified with a the $ sign e.g. edges$with
               Thats gonna change soon.

               The variable on the basis of which the number of edges is selected. E.g, Width or certainty}

\item{cut_off}{If regulation_by = nodes: Scaling factor on how many edges should be selected: E.g. if cut_off = 1.5 and there are 10 nodes
                                         the 15 strongest edges of the cut_by variable are selected

               If regulation_by = number: Number of edges that should be in the network. E.g. if 10 the 10 strongest edges of the
               cut_by variable are selected

               If regulation_by = value: Cut off value for the edges. Edges that have a cut_by vparameter under that value are excluded from the network}

\item{nodes_color}{Color of the nodes. Default is "red"}

\item{edges_color}{Color of the edges. Default  is "black". Can also be "default10" when edges have a width between 1 and 10. "default10"
is a combination of grey and black depending on the width of an edge.}

\item{unweigthed}{Not yet implemented. If TRUE it will set default options for unweigthed networks}

\item{max10}{The possible maximum of your width. Must be numeric.
When your width is not on the scale from 1-10 you can apply this function to scale the maximum of your width to 10.}

\item{arrows}{Where should the arrows be drawn? Can be "to","from","middle" or 'to;from'. Default is "to"}

\item{shadow}{Should some edges have a shadow e.g. to indicate certainty of a specific edge?
A numeric vector indicating the cutoffs for the variable defined by "shadow_by". E.g. c(2,5) --> if an edges has a shadow_by value below or equal to 2 it won´t get a shadow
if it has a shadow_by value lower or equal to 5 it will get a shadow. If it has a value over 5 it will get a thicker shadow.}

\item{shadow_color}{Must be a vector of two colors default is c("#87CEFA","#0000CD").Second color for the thicker edge}

\item{shadow_by}{Variable on which basis of the shadows are applied.}

\item{use_ranking}{Should only edges with a certain rank be used? Must be the column which includes the ranks in edges dataframe. E.g, edges$rank}

\item{min_rank}{Minimum ranking value included egdes should have. e.g min_rank 3 only includes edges which are ranked 3 or below. Care with negative values Must be numeric}

\item{use_seed}{"Default is to random" If u want recreate a network with a certain seed you can set that seed here}
}
\value{
A visNetwork object.
}
\description{
Title
}
