% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_edge_color.R
\name{vis_edge_color}
\alias{vis_edge_color}
\title{Title vis_edge_color}
\usage{
vis_edge_color(net_edges, edges_color, neg_edges_color, edges_color_scaling)
}
\arguments{
\item{edges_color_scaling}{}
}
\description{
Title vis_edge_color
}
