% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nodes_border_color.R
\name{vis_nodes_border_color}
\alias{vis_nodes_border_color}
\title{Title}
\usage{
vis_nodes_border_color(
  net_nodes,
  nodes_border_color,
  nodes_border_color_max,
  nodes_border_color_scaling
)
}
\arguments{
\item{nodes_color_by}{}
}
\description{
Title
}
