% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanExtract.R
\name{pecanExtract}
\alias{pecanExtract}
\title{Title pecanExtract}
\usage{
pecanExtract(
  data,
  p_id,
  edges,
  nodes,
  edges_sep,
  label_in_names = FALSE,
  labels_sep = "none",
  labels = "none"
)
}
\arguments{
\item{data}{df which includes the nodes and edges. Requires a column with the name "p_id"}

\item{p_id}{id of the person of interest...}

\item{edges}{Specify the columns of the edges e.g. 2:12. Edge names need to have a "to" and a "from" specification within their name and a separator.
Names can be numbers E.g 1_2 equals: "from" = 1 and "to" = 2 or words: Pain_Insomnia equals "from" = Pain and "to" = Insomnia}

\item{nodes}{Specify the columns of the nodes e.g. 13:17. Node names must match the edge names!}

\item{edges_sep}{How are "from" and "to" separated in the names?}

\item{label_in_names}{Logical. Are there additional labels in the node names. Must be separated and be second part of the name. e.g 1_Pain or Insomnia_I couldnt sleep}

\item{labels_sep}{How are the labels separated in the node names}

\item{labels}{Its possible to just add vector with the labels}
}
\value{
a list with a node df and an edge df for a specific person
}
\description{
Title pecanExtract
}
