% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanEx.R
\name{pecanExtract}
\alias{pecanExtract}
\title{Title}
\usage{
pecanExtract(
  data,
  p_id,
  edges,
  nodes,
  edges_sep,
  label_in_names = FALSE,
  labels_sep = "none",
  labels = "none"
)
}
\description{
Title
}
