% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanExtract.R
\name{pecanExtract}
\alias{pecanExtract}
\title{Title}
\usage{
pecanExtract(
  data,
  id_col = NULL,
  id_number = NULL,
  edges,
  nodes,
  labels,
  edges_sep,
  nodes_labels = NULL,
  label_pattern = NULL,
  nodes_attributes = NULL,
  attribute_pattern = NULL,
  attribute_name = "attribute",
  edges_direction = "from_to"
)
}
\description{
Title
}
