% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanVis.R
\name{pecanVis}
\alias{pecanVis}
\title{pecanVis}
\usage{
pecanVis(
  nodes,
  edges,
  nodes_size = "auto",
  nodes_size_max = "auto",
  nodes_color = "red",
  nodes_color_by = NULL,
  nodes_color_max = NULL,
  nodes_color_scaling = "scaled",
  nodes_community = NULL,
  nodes_hide_isolated = FALSE,
  nodes_label = "auto",
  nodes_border_color = NULL,
  nodes_border_color_by = NULL,
  nodes_border_color_max = NULL,
  nodes_border_width = NULL,
  nodes_border_color_use_com = TRUE,
  nodes_font = NULL,
  nodes_hidden = NULL,
  edges_color = "black",
  neg_edges_color = "mediumblue",
  edges_color_scaling = "fixed",
  edges_color_by = NULL,
  edges_color_max = NULL,
  edges_regulation = NULL,
  edges_width = "auto",
  edges_width_max = "auto",
  edges_arrows = "to",
  edges_dashes = NULL,
  edges_hidden = "auto",
  edges_label = "auto",
  edges_font = NULL,
  edges_arrowsScaleFactor = 1,
  edges_color.inherit = FALSE,
  edges_selfReference = NULL,
  edges_smooth = TRUE,
  edges_widthConstraint = FALSE,
  edges_arrowStrikethrough = FALSE,
  edges_endPointOffset = NULL,
  use_seed = "random",
  c_width = "100\%",
  c_height = "100\%",
  edit = TRUE,
  export = TRUE,
  auto_detect = TRUE,
  p_id = NULL
)
}
\arguments{
\item{nodes}{nodes data.frame. Must contain an "id" variable for the nodes}

\item{edges}{edges data.frame. Must contain a "from" , a "to"  and still a "width" variable. Plan is that width can also be an other variable soon.}

\item{nodes_color}{Color of the nodes. Default is "red"}

\item{edges_color}{Color of the edges. Default  is "black". Can also be "default10" when edges have a width between 1 and 10. "default10"
is a combination of grey and black depending on the width of an edge.}

\item{edges_hidden}{}

\item{use_seed}{"Default is to random" If u want recreate a network with a certain seed you can set that seed here}
}
\value{
A visNetwork object.
}
\description{
Title
}
